Add files via upload
This script iterates over a folder and finds the length in seconds of all audio files. It creates a dataframe of the data using pandas, and then outputs it to a csv file. Thanks to https://stackoverflow.com/questions/7833807/get-wav-file-length-or-duration for the formula for finding the seconds on a wav file by converting from the bytes.